+++ backend/mockNewsData.js
@@ -787,4 +787,58 @@ const news = [
   },
 ];
 
-module.exports = { news };
+const intervals = {
+  "1-8.5-09": {
+    score: 2.8333333333333335,
+    count: 4,
+    categories: ["entertainment", "general"],
+  },
+  "4-20.5-21": {
+    score: 15,
+    categories: ["entertainment", "general"],
+    count: 1,
+  },
+  "4-18.5-19": {
+    score: 4,
+    categories: ["entertainment", "general"],
+    count: 1,
+  },
+  "1-7.5-08": {
+    score: 4,
+    categories: ["entertainment", "general"],
+    count: 1,
+  },
+  "1-17.5-18": {
+    score: 5,
+    count: 2,
+    categories: ["entertainment", "general"],
+  },
+  "3-7.5-08": {
+    score: 4,
+    categories: ["entertainment", "general"],
+    count: 1,
+  },
+  "3-8.5-09": {
+    score: 1.8333333333333333,
+    count: 4,
+    categories: ["entertainment", "general"],
+  },
+};
+
+const dateScores = {
+  "Thu Jul 17 2025": 0.6666666666666666,
+  "Mon Jul 21 2025": 1.6666666666666665,
+  "Wed Jul 23 2025": 1,
+};
+const categoryScores = { entertainment: 10, general: 10 };
+const sourcesScores = { "www.usmagazine.com": 4.5, "www.eonline.com": 0.5 };
+const sentimentScores = { NEGATIVE: 2, POSITIVE: 3 };
+
+module.exports = {
+  news,
+  intervals,
+  dateScores,
+  categoryScores,
+  sourcesScores,
+  sentimentScores,
+};
